Fix compiler warnings with meson version 0.63.0
Unit Tests:
  1. Enabled below tests in local ci and tests are passed.
    - ssl_key_handler_tests.hpp - covers above functionality
    - file_watcher_tests.hpp - covers InotifyFileWatcher

Fixes jira https://upstream-sync

+++ redfish-core/include/utils/port_utils.hpp
@@ -50,23 +50,23 @@ inline std::string getLinkSpeedGeneration(double speed)
     {
         return "Gen1";
     }
-    if (speed == 5)
+    if (static_cast<int>(speed) == 5)
     {
         return "Gen2";
     }
-    if (speed == 8)
+    if (static_cast<int>(speed) == 8)
     {
         return "Gen3";
     }
-    if (speed == 16)
+    if (static_cast<int>(speed) == 16)
     {
         return "Gen4";
     }
-    if (speed == 32)
+    if (static_cast<int>(speed) == 32)
     {
         return "Gen5";
     }
-    if (speed == 64)
+    if (static_cast<int>(speed) == 64)
     {
         return "Gen6";
     }
